+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 #### New Features
 * Fix #7048: Support for Kubernetes v1.34 (Of Wind & Will)
+* Fix #7174: Added Vert.x 5 HTTP client implementation with improved async handling and WebSocket separation
 
 #### _**Note**_: Breaking changes
 * Fix #7266: bump jackson-bom from 2.19.2 to 2.20.0, fix overrides and handle jackson-annotations v2.20